今までのやつ全部

+++ ArduinoBaseControl/dynamics.cpp
@@ -1,13 +1,13 @@
 #include "dynamics.hpp"
 
-dynamics_t::dynamics_t(double inertia_mass, double viscous_friction_coeff, double torque_constant)
+dynamics_t::dynamics_t(float inertia_mass, float viscous_friction_coeff, float torque_constant)
     : inertia_mass_(inertia_mass),
       viscous_friction_coeff_(viscous_friction_coeff),
       torque_constant_(torque_constant) {
 }
 
-double dynamics_t::calculate_feedforward_control(double target_vel, double target_accel) const {
-    double ff_torque_force;  // フィードフォワードとして計算されるトルク
+float dynamics_t::calculate_feedforward_control(float target_vel, float target_accel) const {
+    float ff_torque_force;  // フィードフォワードとして計算されるトルク
 
     // 慣性項
     ff_torque_force = inertia_mass_ * target_accel;
@@ -18,6 +18,6 @@ double dynamics_t::calculate_feedforward_control(double target_vel, double targe
     return ff_torque_force;
 }
 
-double dynamics_t::convert_to_current_command(double feedforward_torque_force) const {
+float dynamics_t::convert_to_current_command(float feedforward_torque_force) const {
     return feedforward_torque_force / torque_constant_;
 }
